requirements.txt
python-telegram-bot==20.7
requests==2.25.1
selenium==4.16.0
beautifulsoup4==4.12.2